requirements.txt
fontawesomefree==6.4.2
django
django_extensions==3.2.3
Faker
Pillow
cairosvg
svgwrite